修复添加产品库BUG,合并

+++ app/src/main/java/com/miguan/yjy/model/services/Services.java
@@ -361,7 +361,7 @@ public interface Services {
      */
     @GET("?action=queryCosmetics")
     Observable<UserProduct> queryCode(
-            @Query("id") Long brandId,
+            @Query("id") int brandId,
             @Query("number") String number
     );
 
@@ -371,7 +371,7 @@ public interface Services {
     @GET("?action=addRemind")
     Observable<String> addRepository(
             @Query("user_id") int userId,
-            @Query("brand_id") Long brandId,
+            @Query("brand_id") int brandId,
             @Query("brand_name") String brand_name,
             @Query("product") String product,
             @Query("product_img") String img,
